requirements.txt
 flask 
 torch 
 clip 
 faiss-cpu 
 pillow 
 matplotlib 
 numpy 
 requests
 gunicorn